Initial commit
The "Stocks & News" app is a Python script that provides a concise and informative summary of the latest stock prices and trending news articles related to Tesla (TSLA). Leveraging the Alpha Vantage API, the app fetches real-time stock prices for Tesla, displaying the latest value along with the timestamp. Additionally, the app utilizes the News API to retrieve and showcase the top 3 news articles about Tesla, sorted by popularity and published within the last 24 hours.